uplink(feat): Implement Uplink 'config'
Implement the Uplink 'config'.

The method that opens a project but in the Go uplink package belongs to
the 'Config' is left empty and with a TODO to implement it when the
'Project' module is implemented because it's better to implement it
there to not have to make some struct fields public (when keep it them
private provides better encapsulation) nor having to have a method in
project that creates a project from the c-bindings project struct.

The commit also creates some helper functions to be used only in tests.
For this reason the `helpers::test` sub-module is published now at the
crate scope.

Additionally this commit contains a refactoring in some `unsafe` blocks
which kept a variable declaration aside for after assigning its value
inside of them, however, I've realized that is valid in Rust to just
return the value from the `unsafe` block, hence the variable can be
declared and initialized at the same time, making them more idiomatic
Rust.