uplink-sys(deps): Bump uplink-c version to v1.5.0
Bump the version of the Uplink C bindings to the last published release
which is v1.5.0.

Add some metadata to Cargo regarding Uplink C bindings and bump the
crate version.

Add the new entities to implement in `uplink` crate because of the new
additions that the new Uplink-C has and also some new integration test
to implement regarding to them.