uplink-sys(cargo): Fix version before publishing it
In a previous commit the version for Uplink-c was bumped and the version
of the `uplink-sys` crate too, however, the `uplink-sys` version wasn't
properly increased although it wouldn't break anything.

This commit set the proper version taking advantage that so far the new
version has not been published yet in crates.io.