uplink(feat): Implement Uplink 'metadata'
Implement the Uplink 'metadata'.

Implement a helper function to treat C strings as a slice of bytes
because despite metadata entries are String they can contain more than
one NULL byte and they may not terminate with the NULL byte which is a
requirement for using the `std::ffi::CStr`.

Implement two new helper functions to be used only in tests based on two
existing ones, however these ones return values rather than panicking
and the existing ones that panic are based now on these new ones because
they can inspect the returned value for determining the assertion. These
new ones are useful for tests that compare values but they cannot know
what exact value is of a established list due to exposed
non-deterministic behavior (e.g. the order of items that a hash map
contains).