uplink(feat): Implement all Uplink 'options'
Implement all the Uplink 'options' types.

All the options types are used by project methods, so they are
implemented as sub-module of the project module.

The 'options' types are quite minimal and only a few of them have
constructor functions which return a `Result`. These ones have the
constructor to guarantee the integrity of the values of `str` types for
avoiding error in the underlying c-bindings because they contain bytes
not supported by C (i.e. null byte).

Ideally, we'll add test for the constructors of those 'options' types,
but for now they are left over due to their value is minimal in
comparison of other priorities.