uplink(feat): Implement 'object' and its 'iterator'
Implement the Uplink 'object' and its associated iterator. These
implementations required to implement a new static method in the
`System` metadata type for creating an instance from pointer to an
instance of the underlying c-binding system metadata type.

Update the README marking the tasks for implementing these 2 items as
done.