uplink(refactor): Bucket from_uplink_c method
Refactor the `from_uplink_c` `Bucket` static method to consume the
received corresponding c-bindings pointer for not having to hold it in
the created Bucket instance because it isn't needed. This removes the
`Drop` implementation because the pointer isn't hold anymore and release
the memory of the passed c-binding pointer in the same `from_uplink_c`
method.

Also refactor some documentation comments to be more understandable and
fix a few typos and grammar mistakes.