uplink(doc): Format code in access comments
Format some code keywords used in the doc comments of the access module
for improving the readability in the generated HTML documentation pages
and apply some minor fixes in some of them for being less verbose and
more readable.

This commit also refactor all the comments to 100 columns long because
it seems that's the length mostly adopted in Rust sources, rather than
the 80 columns which I thought.