uplink(fix): Fix uplink error details
There was a test for the `access::Grant::override_encryption_key` which
was commented until implementing the `encryption_key` module.

I implemented the `encryption_key` module but I forgot to uncomment the
test. I realized about it and I uncommented the test and added an OK
case, then an unexpected error returned due to an undocumented
constraint for the `prefix` parameter.

Due to the returned error, I've found a bug in the `Error::Uplink` type
regarding in converting the details from C to Rust world.

This commit fix the error's bug and uncomment the test besides of adding
a couple of new cases, an OK case and an error due to a prefix which is
valid according the C rules but invalid because of the constraint of the
argument.