uplink(feat) Implement Uplink `upload`
Implement all the types, and their operations, related with uploading
data to Storj DCS and mark as done all those items listed in the README.

Because there are several types/entities related with the different ways
of uploading data, I've created them as sub-module of the object module.

It's a sub-module of the `object` module because uploading (and also
downloading) data are always associated with an object.