uplink(refactor): Don't return when can never happen
There was a custom metadata constructor that was returning a `Result`
despite it was never returning an error because I thought that it was a
good idea for not breaking changes in the future it needs to return an
error.

Now I thought that I overthought about it and it's kind of
over-engineering it and anyway it wouldn't break changes external
because the function is only visible inside of the crate. Hence, I've
changed it for not returning a `Result`.

I also minimally rephrased a few doc and code comments for making them
easy to read/understand.