uplink(refactor): Error::Uplink from struct to enum
hange the `Uplink` struct for an enum in the error module because it
will allow to the consumers of this crate to identify the error codes
that the Uplink c-bindings return using a `match` expression which is
more Rust idiomatic.

On the other hand, Uplink error codes are a defined set, unless this
crate get out of date because of new Uplink versions add more; In that
case this crate  will return an unknown error and the caller could
report that this crates has a bug or requires an update.

The error modules is now exported publicly for allowing the consumers of
this create to be able to intro-expect the errors because the `Error`
enum use those other types for their variants and those types contains
the error details.

This commit has also readdressed all the code comments of the Error
module to use up to 100 columns rather than 80 because it looks that 100
is the standard used in Rust source code.