uplink(refactor): Rename type & re-export some
Rename the `object::Info` type to `object::Object` and re-export it from
the root of this crate. Although `object::Object` stutters it isn't
something it would be imported from it's real path because it's
re-exported from the crate's root.
Re-exporting it as `crate:Object` type and `crate::object` looks nicer
when they imported because `Info` (the old name) would require the
module path (e.g `use object;` and then `object::Info`) for making
clearer that `Info` holds information of an object.

On the other hand, `bucket::Bucket` already existed, it wasn't
`bucket::Info`, but it wasn't re-exported from the crate's root. This
commit has re-exported this one too.