uplink-sys(deps): Bump uplink-c version to v1.6.1
Bump the uplink-c version to 1.6.1 (the last published version so far).

Arrange the build script for making it work with Go 1.18 but keeping
compatibility with the previous supported Go versions.

Enhance the test to read the passphrase from the secrets file rather
than requiring to be empty because the satellite web UI doesn't
currently allow to set an empty passphrase.