uplink(deps): Bump uplink-sys to v0.3.0
Bump the version of the uplink-sys to v0.3.0 which is the last published
version.

The new version which is based on the Uplink v1.6.1 has 2 new error
codes and 2 new project's methods.

* The new 2 error codes are mapped in this commit.
* This new 2 new corresponding functions for the project's methods
(methods in Go and functions in C and Rust uplink-sys crate) are not
mapped because the project struct type is still not implemented, but,
the passed options to one of the methods is mapped because the options
are already implemented.