uplink(refactor): Don't return error on NULL pointers
Change the `from_uplink_c` constructors that returned an error when the
Uplink-c pointer was NULL for panicking because that should only happen
if there is a bug in the crate implementation due to that they have
crate visibility.

The constructors that returned a Result because of the NULL pointer,
they directly return now the instance. The ones that still return a
Result are because they receive a Uplink-c Result type or they fields
can contain invalid data that prevents to create the instance.