uplink(feat): Implement Edge operations
Implement all the Edge operations and related types.

The implementation has involved a refactoring in the Access Grant due
to some module helper function has been moved to the `uplink_c` module
and it has changed to return a Result of `&str` rather than of `String`.
This change has provoked to change the return type of 2 of the methods
of the `Grant` struct.

I consider that rather that method signature change is a bug fix rather
than a refactoring because I believe that at the time that I created it
I didn't know that I could return a Result of `&str`, despite that's I
would have wanted.