uplink-sys(deps): Bump uplink-c version to v1.7.0
Bump the uplink-c version to 1.7.0 (the last published so far).

We have bumped to this new minor version without updating to the next
patch version because the next patch version only had changes in the
dependencies, nothing from the API was changed, removed, or added.

Bump the crate version to the next minor version.