uplink-sys(fix): Bump version to fix published crate
The last published version of the uplink-sys crate was not correct
because it used an stale git submodule causing that not all the
functions, types, and constants of the target uplink-c version.

Bump the version of the crate for publishing this new version with the
correct git submodule so the crate can expose all the API corresponding
to the bound uplink-c version.