uplink(deps): Bump uplink-sys to v0.5.1
Bump the version of the uplink-sys to v0.5.1 because the v0.5.0 wasn't
published correctly and it breaks the uplink because of missing to
expose some public functions, types, and constants present in the
uplink-c target version.

I forgot to do this before publishing the uplink v0.1.0, hence, the
uplink-v0.1.0 is broken and uplink-v0.1.1 is released with this change.