all(ci): Generalize & homogenize GH Actions workflows
Generalize the repository having a Makefile in the root directory to run
the same target to all the crates that it contains.

Change the uplink workflow for using the root Makefile for building the
uplink-sys crate rather than executing the direct Makefile target of the
uplink-sys crate.

Remove unnecessary verbose comments in the uplink-sys Makefile and
Github workflow and apply some other changes to homogenize them with the
uplink ones.