uplink-sys(build): Allows docs.rs to crate docs
docs.rs couldn't create the docs of uplink-sys crate because the build
image that it uses doesn't have Go installed.

Add the uplink-c binaries libraries for Linux x86_64 architecture, trick
the build script to work differently when executed by docs.rs, and
restrict docs.rs to only build the crate with this architecture works.

Also update the Github workflow for verifying on each PR that the
uplink-c lib binaries are up to date with the version that the
git submodules references for not allowing to merge a PR without having
them updated.