uplink(fix): Make a ctor & member public (#23)
Rust linter shows some warning about functions, methods, and members
visibility.

Not always those warnings apply because it doesn't consider that some
members are read if they are passed to the FFI as raw pointers, however,
it was right for these ones. These ones must be public otherwise the
dependents of this crate cannot use them.

The commit also remove an explicit borrow because the compiler is
already doing it automatically. This was reported as an error by a newer
version of Clippy running on local.