all(release): Bump version to all crates (#24)
Bump the version to all the crates in this repository.

This new `uplink-sys` version includes a workaround for making possible
to build the documentation of the `uplink-sys` by docs.rs, and
consequently, for any crate that depends on it.

The new `uplink` version exposes with public visibility some existing
functions and fields that weren't exposed initially by mistake and
address a Clippy error. It also updates the `uplink-sys` dependency to
the version that this commit bumps.