all(doc): Add missing badges to READMEs
Add badges to all the READMES.
The badges are:

* Github actions status.
* Crates.io latest published version.
* Docs.rs link to the docs of the latest version.
* Crates.io downloads

The main README reference to the `uplink` crate because it's the
idiomatic and safe implementation, hence everybody should use that one.

The README of each specific crate links to its own Github action,
Crates.io, and Docs.io.

Before this only the `uplink-sys` README had Github actions badges and
crates.io.

This commit also changes the style of the badge and mark as done one fo
the tasks listed in the `uplink` README which was missed when it was
completed.