uplink(fix): Make public bucket & object modules
`bucket` and `object` modules had to be public since the beginning
because the contain more than one struct that are needed to be able to
use all the functionalities of the crate.

I didn't likely make them public since the beginning because when I was
developing, I created first the main structs `Bucket` and `Object` and I
reexported them to the crate's root, and after I added the other structs
when I discovered them but I forgot to make the modules public.