uplink(fix): Fix C string references Bucket & Object
I misunderstood how the `to_str` method of the `std::ffi::CStr` type
worked despite I wrote some simplified code for understanding if the
returned `&str` was a copy or a reference to the allocated C string.

In my simplified examples, the `&str` was always valid when I tried to
simulate that it was coming from a variable of another narrow scope,
however when I wanted to use them in an integration test that I was
writing, I found that they bucket's name and object's name were pointing
to garbage because the C string was freed right before returning from
the FFI constructor.

After that, I'm quite confident that my simplified examples worked
because they were strings with a static lifetime.

This commit change them to be of the `String` type because, although, it
has to copy the data from C to Rust, I believe that the performance
impact is minimal and it's more maintainable than having to keep the
uplink-c instance and have to free them when the instances are dropped.
Moreover a `String` type represents better that they own that data.