uplink(fix): Retry on length 0 when downloading
The Uplink-c library may return 0 bytes read when downloading an object,
despite not all the data being read. This usually happens when the object
has more than one segment and the read happens between the segments'
boundaries

Without retrying the crate won't allow downloading the entire object's
data when it has more than one segment.

On the other hand, a reader
(https://doc.rust-lang.org/std/io/trait.Read.html) implementation offers
a great number of useful methods out of the box with only implementing
the `read` method, however, read has only to return 0 bytes read when
there is no more data to read, rather than returning EOF as an error.
This commit identifies when Uplink-c returns the EOF error and returns
no error with the bytes read, that in case of not being 0, Rust will
call it again, and then it will again get the EOF error but with 0 bytes
read.