uplink(release): Bump version due to several fixes
Bump the version of the uplink crate after a few commits that have fixed
some bugs and make the crate usable.

We know that those fixes made it usable because we had our first
integration test that it does a basic full-cycle and without them the
test wouldn't have passed.