uplink(fix): Use a storj-up commit that works
We relied on storj-up latest commit in the main branch however new
recent commits have broken it.

To not break it in our local and CI environments, we are compiling
storj-up in a commit that worked when we started to use it here at the
expense that we'll have to update from time to time.

On the other hand, we have requested to start using tag versions in the
storj-up repository because of this.