uplink(fix): Make share prefixes param optional
Creating an access grant to share not always implies prefixes because a
user may want create an access grant without any path restriction and
only allowing certain operations (read, write, etc.), so, the prefixes
lists must be optional.
This parameter had to be an option since the beginning, so it was a bug
in the crate API design.

We have also added a new constructor to the `SharePrefix` for making
more ergonomic to share an entire bucket.