uplink(fix): Borrow grant when opening project
The `open` constructor of the `Project` struct doesn't need to take
ownership of the `Grant` passed as a parameter. Taking the ownership
makes quite useless to use an access grant that isn't only used for
opening the project, for example, creating new shared ones, etc.,
because once the ownership is transferred it cannot be used anymore.

This was a bug in the API design of this crate.