uplink(fix): Take ownership of C strings
I had a big misunderstanding of the `CStr::to_str` method. I thought
because of not being very experienced with Rust and my wrong playground
implementation, that the method gives a owned `&str` when in reality is
a reference to the C string.

This causes that the function that implemented for getting a Rust string
from a C string was wrong due to returned a reference to a C memory that
is free in the same function, hence, the returned `&str` is invalid.

This commit change the function to return a `String` that's a heap
allocated string rather than a `&str` (how do I miss that point although
I read it manytimes and I understand correctly the diffferences between
the stack and the heap???) and change the return type of all the
functions/methods that rely on it.

Unfortunately, changing the return type of public functions/methods are
breaking changes.