uplink(fix): Don't move vector of shared prefixes
The `share` method of the `Grant` struct was moving (taking ownership)
of the passed vector of `SharePrefix` because it was iterating over a
reference.

This provoked that the `SharePrefix` instances were dropped after the
loop and their were dropped dangling the C strings pointers of the FFI
types that represented them.

This caused to sending garbage data as bucket and prefixes, hence, not
having a correct restricted access grant.