uplink(fix): Result Object has to return an Option
Some Uplink methods that return a `Result<Object>` return in some
situations an OK `Result` without an `Object`, hence, the method that
transform an FFI result object into the Rust type has to be a
`Result<Option<Object>>` to represent those cases.

I identified at least one case, that's when a `Project` is created with
an access grant that has delete permission but it doesn't have read
permission, which provokes the `delete_object` to return an OK result
without an object.

To handle these cases, I had to change the signature of the `Project`
public `delete_object` method which is a breaking change. Apart of that
the non public methods that transform the related FFI types Rust had to
changed and adapt the ones that the return signature doesn't change but
they use these FFI methods.

On the other hand, because for these cases the FFI object result type
can have both pointers (the object & error) to NULL, so I had to delete
its `ensure` trait implementation because it isn't applicable.