uplink(maintenance): Bump storj-up tool to v1 (#35)
storj-up has the first published release v1.0.0. Before we were using
the `main` branch to compile the tool until some commits landed in
`main` and broke the integrations tests because a storj-up functionality
stopped working, hence, we had to hard code a commit hash to compile it.

Now storj-up repository has the first release and tag and it will
publish them from time to time, we can use them rather than hard coding
Git commit hashes.

This commit updates the Makefile to use the v1.0.0 tag and updates the
docker-compose file that storj-up generates because the new version
brings changes.