uplink(fix): Don't take ownership on opts func params
Several functions and methods receive some parameter that represent
operation options.

When I created the crate and defined these methods and functions I
didn't realize that they were taking the ownership because I forgot to
pass a reference of them.

Taking the ownership of them is quite bothersome because they don't
allow to use those options after passed to any of those
methods/functions and it doesn't provided any value due to they don't
require to take ownership.

This commit address this issue passing them as a reference rather than
as a value. Unfortunately, these are breaking changes because they
change the signature of several public methods/functions, hence, they
break the API.