uplink(fix): Make upload Info to own key & id
The upload `Info` struct was not taking ownership of the C strings
created by FFI and this caused to point to garbage data because the FFI
type was free it up.

On the other hand, not owning the values forced to specify a lifetime
for them and the `Iterator` implementation was not working because of
them.

It's more appropriated that `Info` owns these fields than keeping
reference to the FFI because in the end in the Rust world it makes more
sense that it's the owner.

Unfortunately, these are breaking changes.