uplink(release): Bump version due to API fix
Bump the version of the uplink crate after fixing a bug related with the
ownership of C strings created by the FFI.

The fix required to change the type of two public fields of a public
struct, hence, they are breaking changes and we have to bump a minor
version.

The release also include some more integration tests.