uplink(fix): Don't take ownership on access param
Methods of the Edge configuration must not take ownership of the access
grant parameters because they only use them for performing they
operations and they don't retain them.

Taking ownership of them, without requiring them, forces the callers not
to be able to use the access grant for subsequent calls to other methods
or functions that imply them without cloning. Cloning isn't ideal and
harms performance.

This commit changes the parameters to receive a reference for not taking
ownership, but unfortunately, it changes their signature of them, hence
they break the public API.