uplink(doc): Tidy README and TODO comments
All the planned end-to-end tests were finished in version v0.6.0 but the
README wasn't tidied according to them.

This commit tidies the README and also identifies all the `TODO`
comments with an issue for tracking them in the issue tracker for having
a good overview of what has to be done.