uplink-sys(test): Run integration tests in storj-up environment
Previously the `uplink-sys` integration tests had to use a Storj network
running somewhere or a external local environment that people had to set
it up manually.

The `uplink` crate was already using `storj-up` for its integration
tests.

This commit generalize the `storj-up` configuration used in the `uplink`
crate to make available to both crates and adapt the `uplink-sys`
Makefile, CI (Github workflow), and integration tests to use the
`storj-up` environment.

The changes make in the `uplink-sys` integration tests are only for
making easier to adapt it to `storj-up` environment despite, now, it
uses another function to get the access grant, the integration test was
not exactly testing the previous function, so it has not changed the end
goal of it.

There is a minimal change in the CI (Github workflow) of the `uplink`
crate that is out of the scope of this commit, but the changes are
minimal and shouldn't confuse the reviewers. The changes are for adding
a new step to the `check-implementation` job to simulate publishing the
crate in order to verify that any merged PR shouldn't break its version
publishing.