uplink(chore): Bump versions of storj-up services
Bump the versions of the storj-up satellite and edge to match the last
published tag version.

Uncomment an end-to-end test assertion that was commented because the
previous satellite version had a bug and it made the test not pass.

Instruction in one of the Makefile recipes had to be adjusted to work
with the new images.

Also, fix a few typos in test assertion messages.