uplink-sys(build): Use Go version to 1.21 for docs
Use the last published stable Go version to compile the library binaries
of the uplink-c used for docs.rs.

Because of security measure applied by default from certain version of
Git, we have to use the local user inside of the container to allow Go
to figure out the version information. Hence, the Makefile has been
adjusted accordingly.