uplink(deps): Bump uplink-sys version
Bump the uplink-sys version to the latest published one.

It has required to adapt some values to uszie because before they were
u64, hence, we don't need any casting.

Because the new version brings an error code that was not mapped because
of a bug in the past, we had to adjust the test to check for this new
error code rather than being an internal error, which is used when it's
an internal one, but also if an error code doesn't match any of the
registered ones.