uplink(refactor): Remove unused fields on structs
Remove some unused fields in some options structs.

I added them because I thought they were going to be convenient if at
some point we want to expose them to the end users, so they can access
to those values without having to deal with CString types.

However, the linter reported them as not used, and I' convinced to
remove them and add them in the future if they are needed for what I
thought it.