all(test): Bump storj-up version
Bump the storj-up version to the latest published release.

After the bump we identified that to a certain point a satellite version
started to use a difference nonce for every new access grant creation,
so the test that creates an access grant from the API and encryption key
will never return the same access grant, despite they will allow to
perform the same access level. Hence I remove one assertion.

We also identified that the error that we got with the Edge::Config for
registering a gateway access was not cause because of using an older
storj-up version, so I rephrased the TODO comment.