all(test): Update storj-up to v1.2.7
Update storj-up to v1.2.7 to use a more recent satellite and edge
versions.

This commit also removes the docker-compose.yaml file to always
autogenerate through storj-up to not have to keep it manually in sync
every time that storj-up is updated.

Thanks to @dlamarmorgan for the help of bumping storj-up and the tip not
to have to autogenerate the docker-compose.yaml file.