all(deps): Bump uplink-c version to v1.9.0
Bump the uplink-c version to the last published release v1.9.0 (minor
release) and bindgen.

Bump the uplink-sys crate version to a new minor release to reflect
these dependencies updates.

Update uplink crate to use the new uplink-sys version and bump it to a
new minor release to reflect this dependency update.